+++ maml/describer/_structure.py
@@ -9,7 +9,7 @@ from pymatgen.symmetry.analyzer import SpacegroupAnalyzer
 from pymatgen.core import Structure
 
 from maml import BaseDescriber
-from ._megnet import MEGNetStructure
+from .megnet import MEGNetStructure
 
 
 __all__ = ['DistinctSiteProperty', 'MEGNetStructure', 'CoulombMatrix',
