+++ maml/describer/matminer/__init__.py
@@ -1,0 +1,6 @@
+"""Matminer related classes"""
+from ._matminer_wrapper import wrap_matminer_describer  # noqa
+
+__all__ = [
+    "wrap_matminer_describer"
+]
