+++ maml/describer/megnet/__init__.py
@@ -1,0 +1,9 @@
+"""
+megnet related describers
+"""
+from ._megnet import MEGNetSite, MEGNetStructure  # noqa
+
+__all__ = [
+    "MEGNetSite",
+    "MEGNetStructure"
+]
