+++ maml/describer/megnet/_megnet.py
