+++ maml/describer/rdf/__init__.py
@@ -1,0 +1,9 @@
+"""
+Radial distribution function related describers
+"""
+
+from ._rdf import RadialDistributionFunction  # noqa
+
+__all__ = [
+    "RadialDistributionFunction"
+]
